+++ pyproject.toml
@@ -24,6 +24,7 @@ dev = [
     "rich>=14.0.0",
     "simple-parsing>=0.1.7",
     "pillow>=11.2.1",
+    "pytest-mock>=3.14.1",
 ]
 
 [tool.pytest.ini_options]
